Moved TriblerApp class to seperate file
